requirements.txt
Flask
scikit-learn
pandas
nltk
numpy
spacy
pdfplumber
python-docx
beautifulsoup4
requests
